requirements.txt
django==4.0.3
djangorestframework
python==3.8.5
corsheaders
djangorestframework-simplejwt